+++ client/src/components/Pages/index.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Route } from "react-router-dom";
 
 import LandingPage from "./LandingPage";
-import InterCreateProfile from "./InternCreateProfile";
+import HostCreateProfile from "./HostCreateProfile";
 
 import { HOME_URL, COMPLETE_PROFILE_URL } from "./../../constants/navRoutes";
 
@@ -13,7 +13,7 @@ class Pages extends Component {
         <Route path={HOME_URL} component={LandingPage} exact />
         <Route
           path={COMPLETE_PROFILE_URL}
-          component={InterCreateProfile}
+          component={HostCreateProfile}
           exact
         />
       </>
